pyproject.toml
[tool.poetry]
name = "digitaldome"
version = "0.1.0"
description = ""
authors = ["Kamil Marut <kamil@kamilmarut.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
django = "^4.2.3"
django-filter = "^23.3"
psycopg = "^3.1.9"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
pillow = "^10.0.0"
pyjwt = "^2.8.0"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.286"
ipython = "^8.14.0"
djlint = "^1.33.0"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
select = ["E", "F", "ANN", "ASYNC", "S", "A", "DJ", "G", "PIE", "T20", "RSE", "SIM", "PTH", "PD", "TRY", "RUF"]
ignore = ["ANN401", "TRY003"]
line-length = 120
target-version = "py311"

[tool.ruff.per-file-ignores]
"**/management/commands/*.py" = ["A003"]
"**/migrations/0*_*.py" = ["RUF012"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
